Clean up class; Expose internals to test; get rid of string result type
