+++ snippets5000/LocateProjects/LocateProjects.csproj
@@ -3,8 +3,9 @@
   <PropertyGroup>
     <OutputType>Exe</OutputType>
     <TargetFramework>net7.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
-	<ImplicitUsings>enable</ImplicitUsings>
+
     <Version>2.0.0.0</Version>
 
     <PackAsTool>true</PackAsTool>
